refactor: move more functions around
