+++ src/stormlibpp/utils.py
@@ -1,7 +1,9 @@
 """Miscellaneous functions that are helpful for working with Storm/Synapse."""
 
 
+import csv
 import getpass
+import json
 import os
 
 import synapse.exc as s_exc
@@ -44,6 +46,67 @@ def chk_storm_syntax(storm: str) -> None:
         raise errors.StormSyntaxError("Storm Syntax Error!", err) from err
 
 
+def csv_genr(fd: str) -> list:
+    """Read a CSV from fd and yield each row."""
+
+    for row in csv.reader(fd):
+        yield row
+
+
+def endswith(text: str, items: list[str]) -> bool:
+    """Check if text ends with any of items."""
+
+    for item in items:
+        if text.endswith(item):
+            return True
+    return False
+
+
+def get_cortex_creds(_user: str | None = None) -> tuple[str, str]:
+    """Get credentials to use when connecting to a Cortex over HTTP.
+    
+    If ``_user`` is set, it overrides any other options to get a username.
+    Otherwise use the value of the ``CORTEX_USER`` environment variable. If that
+    is also empty, prompt for a username. A default option is given to the user
+    at the prompt, it is the return of ``getpass.getuser()``.
+
+    The password is read from the ``CORTEX_PASS`` environment variable and the
+    user is prompted if this is empty.
+    """
+
+    if _user:
+        username = _user
+    elif envusr := os.environ.get("CORTEX_USER"):
+        username = envusr
+    else:
+        gpusr = getpass.getuser()
+        inusr = input(f"Username [{gpusr}]: ")
+        username = inusr if inusr else gpusr
+
+    if envpw := os.environ.get("CORTEX_PASS"):
+        password = envpw
+    else:
+        password = getpass.getpass()
+
+    return username, password
+
+
+def json_genr(fd):
+    """Read a JSON file from fd, and yield each top-level item within it."""
+
+    # TODO - Support nested key to start with
+    data = json.load(fd)
+
+    if type(data, dict):
+        for key, row in data.items():
+            yield [key, row]
+    else:
+        count = 0
+        for row in data:
+            yield [count, row]
+            count += 1
+
+
 def normver(ver: str | tuple) -> tuple[str, tuple]:
     """Take either a version str "x.x.x" or tuple (x, x, x) and return both.
 
@@ -65,19 +128,8 @@ def normver(ver: str | tuple) -> tuple[str, tuple]:
     return (verstr, vertup)
 
 
-def get_cortex_creds(_user):
-    if _user:
-        username = _user
-    elif envusr := os.environ.get("CORTEX_USER"):
-        username = envusr
-    else:
-        gpusr = getpass.getuser()
-        inusr = input(f"Username [{gpusr}]: ")
-        username = inusr if inusr else gpusr
+def txt_genr(fd):
+    """Read a raw text file from fd and yield each line."""
 
-    if envpw := os.environ.get("CORTEX_PASS"):
-        password = envpw
-    else:
-        password = getpass.getpass()
-
-    return username, password
+    for row in fd.readlines():
+        yield row.strip()
